requirements.txt
pymongo
streamlit
litellm
python-dotenv